refactor: 인증 관련 mutation 함수의 추상화 해제
