--- frontend/src/api/user/hooks/index.ts
@@ -1,5 +1,0 @@
-import useCreateUserMutation from './useCreateUserMutation';
-import useGetUsersQuery from './useGetUsersQuery';
-import useLoginMutation from './useLoginMutation';
-
-export { useCreateUserMutation, useGetUsersQuery, useLoginMutation };
