+++ frontend/src/api/user/index.ts
@@ -1,14 +1,14 @@
 import { CreateUserRequestBody, LoginRequestBody } from './type';
 import instance from './instance';
 
-export const createUser = async (body: CreateUserRequestBody) => {
+export const requestCreateUser = async (body: CreateUserRequestBody) => {
   return await instance.post('', body);
 };
 
-export const getUsers = async () => {
+export const requestGetUsers = async () => {
   return await instance.get('');
 };
 
-export const login = async (body: LoginRequestBody) => {
+export const requestLogin = async (body: LoginRequestBody) => {
   return await instance.post('/login', body);
 };
