+++ frontend/src/components/view/sign-up/SignUpForm.tsx
@@ -1,5 +1,9 @@
 import { useForm } from 'react-hook-form';
-import { Email } from '../../../api/user/type';
+import {
+  CreateUserRequestBody,
+  CreateUserResponseBody,
+  Email,
+} from '../../../api/user/type';
 import AuthForm from '../../ui/auth/AuthForm';
 import InputLabel from '../../ui/InputLabel';
 import Input from '../../ui/Input';
@@ -8,16 +12,29 @@ import InputErrorMessage from '../../ui/InputErrorMessage';
 import AuthButtonContainer from '../../ui/auth/AuthButtonContainer';
 import AuthButton from '../../ui/auth/AuthButton';
 import { useNavigate } from '@tanstack/react-router';
-import useCreateUserMutation from '../../../api/user/hooks/useCreateUserMutation';
 import { toast } from 'react-toastify';
+import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { AxiosError, AxiosResponse } from 'axios';
+import { requestCreateUser } from '../../../api/user';
+import { queryKey } from '../../../api/user/queryKey';
 
 type FormInput = {
   email: Email;
 };
 
 export default function SignUpForm() {
+  const queryClient = useQueryClient();
+  const createUserMutation = useMutation<
+    AxiosResponse<CreateUserResponseBody>,
+    AxiosError,
+    CreateUserRequestBody
+  >({
+    mutationFn: (body) => requestCreateUser(body),
+    onSuccess: () => {
+      queryClient.invalidateQueries({ queryKey: queryKey.signup() });
+    },
+  });
   const navigate = useNavigate();
-  const createUserMutation = useCreateUserMutation();
   const {
     formState: { errors },
     register,
